beğeni ve yorumlarla ilgili kısımlar bitti şimdi makale kaydetme daha sonra da makaleleri kategorilere bölmeyi yapacağız
